requirements.txt
groq>=0.9.0
google-generativeai>=0.7.0
chromadb>=0.5.0
python-dotenv>=1.0.0
streamlit>=1.35.0
pydantic>=2.7.0
requests>=2.31.0
langsmith>=0.1.0
pandas>=2.0.0